+++ BranchSDK/src/BranchIO/APIClientSession.cpp
@@ -16,7 +16,7 @@
 
 #include "BranchIO/Defines.h"
 #include "BranchIO/IRequestCallback.h"
-#include "BranchIO/Log.h"
+#include "BranchIO/Util/Log.h"
 
 using namespace std;
 using namespace Poco;
