+++ BranchSDK/src/BranchIO/Util/Log.h
